fixed problem with authfile. now one authfile

+++ reset.sh
@@ -2,11 +2,17 @@
 
 echo "Reset Database ? (WARN:Irreversible action !)"
 read yesno
-if [ $yesno = "y" ] || [ $yesno = "yes" ]
+if [ -z $yesno ]
 then
-	rm -fv *.db tokens/init.token
-	rm -fv configs/mosquitto.conf
-	rm -fv configs/mosquitto.auth
-	rm -fv logs/*.log logs/*.tmp
+	exit 0
+else
+	if [ $yesno = "y" ] || [ $yesno = "yes" ]
+	then
+		echo "Resetting server"
+		rm -fv *.db tokens/init.token
+		rm -fv configs/mosquitto.conf
+		rm -fv configs/mosquitto.auth
+		rm -fv logs/*.log logs/*.tmp
+	fi
 fi
